Eliminated old JsonLocatorField
